Equation 19 of the paper

Equation 19 of the paper

eq_19(x, G, p_n, N = 50L, K = 50L, alpha, theta, parallel = FALSE, ...)

Arguments

x
Support of the probability density function.

G
Baseline cumulative distribution function.

p_n
Zero-truncated power series distribution.

N
Maximum number of sums of the external sum of Eq. 19 of the paper.

K
Maximum number of sums of the interior sum of Eq. 19 of the paper.

alpha
Parameter added by the proposed family.

theta
Parameter added by the proposed family.

parallel
If parallel = TRUE, the code will be executed in parallel, taking advantage of the predecessor's multicolors, if any. If parallel = FALSE, the code will be executed serially.

...
Baseline distribution parameters.

Examples

# Implementing the Weibull distribution function (baseline). This function is
# used in the eq_19 function. Another baseline G can be passed and tested in
# eq_19.
cdf_w <- function(x, beta, lambda) {
   1 - exp(-(lambda * x)^beta)
}

pdf_aprox <- function(x, alpha, theta, lambda, beta) {
  eq_19(
    x = x,
    G = cdf_w,
    p_n = pf_ztp,
    N = 200L,
    K = 100L,
    alpha = alpha,
    theta = theta,
    beta = beta,
    lambda = lambda
  )
}

# True parameters
x = 1.28
alpha = 1.2
theta = 1.5
beta = 1.33
lambda = 2

# Checking theorical density and approximate density for a finite sum of Eq 19 of the paper
pdf_aprox(
   x = x,
   alpha = alpha,
   theta = theta,
   lambda = lambda,
   beta = beta
)

#> [1] 0.2395632

pdf_theorical(
   x = x,
   alpha = alpha,
   theta = theta,
   beta = beta,
   lambda = lambda
)

#> [1] 0.2395632